Move equalizer tag to the heading of the card
It’s the difference in lines taken up by the heading that determine the total height of the card. The images will always remain the same, the height of the prose card is affected the the lines taken by the heading. So put the equalizer tag on the heading since the goal is the button link at the bottom